requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pytesseract==0.3.10
Pillow==10.1.0
